Remove cross domain redirects
Instead inline the content generation. This is correct in more
situations when the short domain name fragments (by default go and
goto) and used as part of multiple different domain names, eg:
goto.subdomain.example.com and goto.example.com. Redirecting across
domains in this situation won't always result in good behaviour.

Ideally this could have been done by mutating the request in place.
Alternatively it could be done by something in front of the app to
rewrite the URL when the domain starts with the quick name, by adding
the '/to' to the beginning of the path.